Remove unnecessary dependencies
The graphite module currently installs a bunch of dependencies
that are a relic of the days when graphite was installed by
pip instead of packages.  Now that we use packages, we should
only declare resources for the packages themselves and let
apt-get take care of dependencies.  Note that x11-apps is left
in as it provides some added (but not essential and therefore
not included as a package dependency) functionality.

Closes-Bug: #1272619